Send access token as httponly cookie and refresh token not at all
"api/token" takes a set of user credentials and returns an access token as httponly cookie. Its nice that only the BE has to deal with JWT tokens if we store the access token in a cookie. As we do not need the refresh token at the moment, we do not send it any more. If we want to get this functionality, we can reactivate it easily.